+++ savintsev.ilya/T1/concave.hpp
@@ -12,8 +12,8 @@ namespace savintsev
     rectangle_t getFrameRect() const override;
     void move(point_t p) override;
     void move(double x, double y) override;
-    void scale(double k) override;
   private:
+  void doScale(double k) override;
     point_t p1_;
     point_t p2_;
     point_t p3_;
