+++ savintsev.ilya/T1/lins-tris.cpp
@@ -1,4 +1,4 @@
-#include "on-plane.hpp"
+#include "lins-tris.hpp"
 #include <cmath>
 
 double savintsev::getLength(point_t lhs, point_t rhs)
